Create ServerConfig to group server configuration in ServerContext
